raise an exception if the response was bad
Unfortunately it appears as though the OpenStreetMaps organization has
started blocking requests made from python scripts :(

As someone who has internet accessible servers, I completely understand
the rationale behind this decision and why I will not implement a
workaround for this.

This commit simply updates these scripts to return the status code of
the response when it is block by OpenStreetMaps.

Signed-off-by: Bryan Brattlof <bb@ti.com>